remove package-lock.json and move getUnderlyingPrice
